requirements.txt
South
Django==1.5.7
MySQL-python==1.2.4
gevent==0.13.8
greenlet==0.4.1
gunicorn==18.0
psycopg2==2.5.1
python-dateutil==2.1
requests==2.0.0
zope.dottedname==4.0.1

djangohelpers
django-debug-toolbar==0.9.4
django-dbtemplates
dj-database-url==0.2.1

git+git://github.com/ejucovy/django-oauth2-provider.git#egg=django-oauth2-provider
git+git://github.com/boldprogressives/django-actionkit-client.git#egg=django-actionkit
